Updated Data Types

+++ inc/core/internal/vt_database.h
@@ -16,16 +16,16 @@ void _vt_database_initialize_falltimedb(VT_DATABASE* database_ptr);
 void _vt_database_initialize_pearsoncoefficientdb(VT_DATABASE* database_ptr);
 
 // Store
-uint32_t _vt_database_store_falltime(VT_DATABASE* database_ptr, uint32_t fall_time, int sensor_id);
+uint32_t _vt_database_store_falltime(VT_DATABASE* database_ptr, uint32_t fall_time, uint8_t sensor_id);
 
-uint32_t _vt_database_store_pearsoncoefficient(VT_DATABASE* database_ptr, float pearson_coefficient, int sensor_id);
+uint32_t _vt_database_store_pearsoncoefficient(VT_DATABASE* database_ptr, float pearson_coefficient, uint8_t sensor_id);
 
 //Check if Empty
-int _vt_database_check_pearson_falltime_availability(VT_DATABASE* database_ptr);
+uint32_t _vt_database_check_pearson_falltime_availability(VT_DATABASE* database_ptr);
 
 // Evaluate
-int _vt_database_evaluate_nrmse(VT_DATABASE* database_ptr, uint32_t* fallcurvearray);
+int8_t _vt_database_evaluate_nrmse(VT_DATABASE* database_ptr, uint32_t* fallcurvearray);
 
-int _vt_database_evaluate_pearson_falltime(VT_DATABASE* database_ptr, int fall_time, float pearson_coefficient);
+int8_t _vt_database_evaluate_pearson_falltime(VT_DATABASE* database_ptr, uint32_t fall_time, float pearson_coefficient);
 
 #endif
