+++ src/core/vt_database_fetch.c
@@ -5,14 +5,14 @@
 
 #include "vt_database.h"
 
-uint32_t vt_database_falltime_fetch(VT_DATABASE* database_ptr, int* index, int* fall_time, int* sensor_id)
+uint32_t vt_database_falltime_fetch(VT_DATABASE* database_ptr, uint8_t* index, uint32_t* fall_time, uint8_t* sensor_id)
 {
     if (database_ptr == NULL)
     {
         return (VT_PTR_ERROR);
     }
 
-    if (!(*index < (int)(database_ptr->_vt_total_falltime)))
+    if (!(*index < database_ptr->_vt_total_falltime))
     {
         *index = -1;
         return (VT_SUCCESS);
@@ -25,26 +25,26 @@ uint32_t vt_database_falltime_fetch(VT_DATABASE* database_ptr, int* index, int* 
 }
 
 uint32_t vt_database_pearsoncoefficient_fetch(
-    VT_DATABASE* database_ptr, int* index, float* pearson_coefficient, int* sensor_id)
+    VT_DATABASE* database_ptr, uint8_t* index, float* pearson_coefficient, uint8_t* sensor_id)
 {
     if (database_ptr == NULL)
     {
         return (VT_PTR_ERROR);
     }
 
-    if (!(*index < (int)(database_ptr->_vt_total_pearson_coefficient)))
+    if (!(*index < database_ptr->_vt_total_pearson_coefficient))
     {
         *index = -1;
         return (VT_SUCCESS);
     }
 
-    *sensor_id            = (int)(database_ptr->_vt_pearson_coefficientdb[*index][0]);
-    *pearson_coefficient = database_ptr->_vt_pearson_coefficientdb[*index][1];
+    *sensor_id            = database_ptr->_vt_pearson_coefficientdb[*index][0];
+    *pearson_coefficient  = database_ptr->_vt_pearson_coefficientdb[*index][1];
     (*index)++;
     return VT_SUCCESS;
 }
 
-int _vt_database_check_pearson_falltime_availability(VT_DATABASE* database_ptr)
+uint32_t _vt_database_check_pearson_falltime_availability(VT_DATABASE* database_ptr)
 {
     if (database_ptr->_vt_total_falltime == 0)
     {
